fix(split-chunks): align rebased cache group indices

+++ crates/rspack_plugin_split_chunks/src/plugin/module_group.rs
@@ -134,7 +134,7 @@ impl AffectedModuleGroupIndex {
 }
 
 struct ModuleGroupHeapEntry {
-  cache_group_index: usize,
+  cache_group_index: u32,
   chunks_len: usize,
   key: String,
   modules: Vec<ModuleIdentifier>,
@@ -923,7 +923,7 @@ impl SplitChunksPlugin {
 #[allow(clippy::too_many_arguments)]
 fn process_chunk_combination_fast_path(
   module_identifier: ModuleIdentifier,
-  cache_group_index: usize,
+  cache_group_index: u32,
   cache_group: &CacheGroup,
   chunk_combination: &FxHashSet<ChunkUkey>,
   removed_chunks: Option<&FxHashSet<ChunkUkey>>,
@@ -1003,7 +1003,7 @@ fn merge_module_group_maps(left: &mut ModuleGroupMap, right: ModuleGroupMap) {
 async fn process_chunk_combination(
   module: &dyn Module,
   module_identifier: ModuleIdentifier,
-  cache_group_index: usize,
+  cache_group_index: u32,
   cache_group: &CacheGroup,
   chunk_combination: &FxHashSet<ChunkUkey>,
   removed_chunks: Option<&FxHashSet<ChunkUkey>>,
@@ -1141,7 +1141,7 @@ async fn merge_matched_item_into_module_group_map(
 
 fn merge_matched_item_into_module_group_map_sync(
   module_identifier: ModuleIdentifier,
-  cache_group_index: usize,
+  cache_group_index: u32,
   cache_group: &CacheGroup,
   selected_chunks: &[ChunkUkey],
   module_group_map: &mut ModuleGroupMap,
@@ -1175,7 +1175,7 @@ fn merge_matched_item_into_module_group_map_sync(
 
 fn merge_matched_item_into_module_group_map_with_set(
   module_identifier: ModuleIdentifier,
-  cache_group_index: usize,
+  cache_group_index: u32,
   cache_group: &CacheGroup,
   selected_chunks: &FxHashSet<ChunkUkey>,
   module_group_map: &mut ModuleGroupMap,
@@ -1209,7 +1209,7 @@ fn merge_matched_item_into_module_group_map_with_set(
 
 fn merge_matched_item_into_module_group_map_with_set_static_name(
   module_identifier: ModuleIdentifier,
-  cache_group_index: usize,
+  cache_group_index: u32,
   cache_group: &CacheGroup,
   selected_chunks: &FxHashSet<ChunkUkey>,
   module_group_map: &DashMap<String, ModuleGroup>,
